split register task into 2, one as part of the form and another as background task
