+++ opwen_email_client/webapp/actions.py
@@ -1,9 +1,7 @@
 from contextlib import contextmanager
 from importlib import import_module
-from json import dumps
 from json import loads
 from logging import Logger
-from logging import getLogger
 from pathlib import Path
 from requests.exceptions import HTTPError
 from requests import get
@@ -30,7 +28,9 @@ from opwen_email_client.domain.modem import modem_is_setup
 from opwen_email_client.domain.modem import setup_modem
 from opwen_email_client.domain.sim import dialup
 from opwen_email_client.util.os import backup
+from opwen_email_client.webapp.config import AppConfig
 from opwen_email_client.webapp.config import i8n
+from opwen_email_client.webapp.config import root_domain
 
 
 class SyncEmails(object):
@@ -232,39 +232,51 @@ class StartInternetConnection(object):
 
 
 class ClientRegister(object):
-    def __init__(self, client_name: str, username: str, access_token: str, logger: Logger):
+    def __init__(self, client_name: str, access_token: str, path: str, logger: Logger):
         self._client_name = client_name
-        self._github_username = username
         self._github_access_token = access_token
-        self._log = logger or getLogger(__name__)
+        self._settings_path = path
+        self._log = logger
 
     @property
     def client_domain(self):
         return '{}.{}'.format(self._client_name, 'lokole.ca')
 
     @property
-    def client_url_create(self):
-        return 'https://{}/api/email/register/'.format('mailserver.lokole.ca')
-
-    @property
     def client_url_details(self):
-        return 'https://{}/api/email/register/{}'.format('mailserver.lokole.ca', self.client_domain)
+        return 'https://{}/api/email/register/{}'.format('mailserver.lokole.ca', self.client_domain)\
+
+    def _fetch_settings(self):
+        if AppConfig.RESTART_PATHS:
+            restart_paths_list = ['{}={}'.format(key, value) for (key, value) in
+                                    AppConfig.RESTART_PATHS.items()]
+            restart_path = ','.join(restart_paths_list)
+        else:
+            restart_path = ''
+
+        return {
+            'OPWEN_APP_ROOT': AppConfig.APP_ROOT,
+            'OPWEN_STATE_DIRECTORY': AppConfig.STATE_BASEDIR,
+            'OPWEN_SESSION_KEY': AppConfig.SECRET_KEY,
+            'OPWEN_MAX_UPLOAD_SIZE_MB': AppConfig.MAX_UPLOAD_SIZE_MB,
+            'OPWEN_SIM_TYPE': AppConfig.SIM_TYPE,
+            'OPWEN_EMAIL_SERVER_HOSTNAME': AppConfig.EMAIL_SERVER_HOSTNAME,
+            'OPWEN_CLIENT_NAME': self.client_name.data.strip(),
+            'OPWEN_ROOT_DOMAIN': root_domain,
+            'OPWEN_RESTART_PATH': restart_path,
+        }
+
+    def _write_settings_to_file(self, client_settings, client_details):
+        client_settings_list = ['{}={}'.format(key, value) for (key, value) in
+                                client_settings.items()]
+        client_details_list = ['{}={}'.format(key, value) for (key, value) in
+                                client_details.items()]
+
+        with open(self._path, 'w') as fobj:
+            fobj.write('\n'.join(client_settings_list))
+            fobj.write('\n'.join(client_details_list))
 
     def __call__(self):
-        create_request_payload = dumps({'domain': self.client_domain}).encode('utf-8')
-        create_headers = {'Content-Type': 'application/json',
-                   'Content-Length': str(len(create_request_payload)),
-                   'Authorization': 'Bearer {}'.format(self._github_access_token)}
-
-        response = post(self.client_url_create, data=create_request_payload, headers=create_headers)
-        try:
-            response.raise_for_status()
-        except HTTPError as ex:
-            self._log.exception('Unable to register client {client_name}: [{status_code}] {message}'.format(
-                client_name=self._client_name,
-                status_code=ex.code,
-                message=ex.read().decode('utf-8').strip()
-            ))
 
         while(True):
             get_headers = {'Authorization': 'Bearer {}'.format(self._github_access_token)}
@@ -285,9 +297,12 @@ class ClientRegister(object):
                 client_info = loads(get_response.text)
                 break
 
-        return {
+        opwen_settings = self._fetch_settings()
+        registration_details = {
             'OPWEN_CLIENT_ID': client_info['client_id'],
             'OPWEN_REMOTE_ACCOUNT_NAME': client_info['storage_account'],
             'OPWEN_REMOTE_ACCOUNT_KEY': client_info['storage_key'],
             'OPWEN_REMOTE_RESOURCE_CONTAINER': client_info['resource_container'],
         }
+
+        self._write_settings_to_file(opwen_settings, registration_details)
